Moved CommandVariables to their own file
